ecom_dataset/requirements.txt
pandas
scikit-learn
matplotlib
seaborn
scipy
xgboost
lightgbm
catboost
statsmodels
imblearn
mlflow
holidays
category-encoders
ydata-profiling
geopy
flaml
python-dotenv